requirements.txt
streamlit
yfinance
pandas
prophet
finnhub-python
newsapi-python
python-dotenv
numpy
matplotlib
plotly
textblob

